requirements.txt
asgiref==3.10.0
astroid==4.0.2
attrs==25.4.0
dill==0.4.0
Django==5.0.6
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
graphviz==0.21
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.9
pydotplus==2.0.2
PyJWT==2.9.0
pylint==4.0.4
pyparsing==3.2.5
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
tomlkit==0.13.3
typing_extensions==4.15.0
uritemplate==4.2.0
